aider: The script now processes files in the `git/template` directory with `mustache` using the provided JSON file, outputting the results to a new directory within `temp`.
# Aider chat conversation:

USER: Please, make generate.sh take a single parameter, name of json file. When it's run, it should, using the mustache command line, create a folder within temp that contains the contents of template in this git directory with all the files having been run through mustache. I have the command line version installed using the json file for data.

ASSISTANT: ```diff
--- tmp/generate.sh
+++ tmp/generate.sh
@@ ... @@
 #! /bin/bash
+if [ "$#" -ne 1 ]; then
+    echo "Usage: $0 <json_file>"
+    exit 1
+fi
+
+JSON_FILE=$1
+TEMPLATE_DIR="git/template"
+OUTPUT_DIR="temp/$(basename "$JSON_FILE" .json)"
+
+mkdir -p "$OUTPUT_DIR"
+cp -r "$TEMPLATE_DIR/"* "$OUTPUT_DIR/"
+
+find "$OUTPUT_DIR" -type f -exec sh -c 'mustache "$1" "$2" > "$2".tmp && mv "$2".tmp "$2"' _ "$JSON_FILE" {} \;
```

This script now takes a single parameter, the name of the JSON file, and uses the `mustache` command line tool to process the files in the `git/template` directory with the data from the JSON file, outputting the results to a new directory within `temp`.