aider: Modified script to check for existing and non-empty directory before copying contents, and updated find command to use a while read loop for processing files.
# Aider chat conversation:

USER: I ran this command:

 ./tmp/generate.sh data.json

And got this output:

cp: cannot stat 'git/template/*': No such file or directory

ASSISTANT: It seems like the script is unable to find the `git/template` directory or the directory is empty. To address this issue, I will modify the script to check if the `TEMPLATE_DIR` exists and is not empty before attempting to copy its contents. If the directory does not exist or is empty, the script will output an appropriate error message.

Here is the diff for the updated `generate
^C KeyboardInterrupt

USER: Error: Unclosed tag at 15413
    at parseTemplate (/home/kian/.bun/install/global/node_modules/mustache/mustache.js:217:15)
    at Writer.parse (/home/kian/.bun/install/global/node_modules/mustache/mustache.js:523:16)
    at Writer.render (/home/kian/.bun/install/global/node_modules/mustache/mustache.js:554:23)
    at Object.render (/home/kian/.bun/install/global/node_modules/mustache/mustache.js:758:26)
    at render (/home/kian/.bun/install/global/node_modules/mustache/bin/mustache:101:15)
    at invoke (/home/kian/.bun/install/global/node_modules/mustache/bin/mustache:49:8)
    at invokeNextFn (/home/kian/.bun/install/global/node_modules/mustache/bin/mustache:45:5)
    at ReadStream.onEnd (/home/kian/.bun/install/global/node_modules/mustache/bin/mustache:118:5)
    at Object.onceWrapper (node:events:633:28)
    at ReadStream.emit (node:events:519:28)

Node.js v21.7.1

ASSISTANT: The error message you're seeing indicates that there is an unclosed tag in one of the template files being processed by Mustache. This is a problem with the template
^C KeyboardInterrupt

USER: Please change the last line of the find command to pipe in the list of files to a while read loop and perform the actions within the loop.

ASSISTANT: ```diff
--- tmp/generate.sh
+++ tmp/generate.sh
@@ ... @@
-find "$OUTPUT_DIR" -type f -exec sh -c 'mustache "$" "$2" > "$2".tmp && mv "$2".tmp "$2"' _ "$JSON_FILE" {} \;
+find "$OUTPUT_DIR" -type f | while read -r file; do
+    mustache "$JSON_FILE" "$file" > "$file".tmp && mv "$file".tmp "$file"
+done
```

This change replaces the `-exec` option of the `find` command with a `while read` loop that reads each file found and applies the `mustache` command to it within the loop.